requirement.txt
fastapi
uvicorn
requests
web3
python-dotenv
opencv-python